Move all URL related stuff from AppUtils to URLUtils class (#425)
